chore: extract kazflamethrower.h definitions of core2 functions into core2 specific header files, re-use where applicable

+++ src/overlays/bskazfly/bskazfly.c
@@ -299,7 +299,7 @@ void func_80800BDC_bskazfly(PlayerState *self) {
     f32 unk0;
     f32 unk1[3];
     s32 pad2;
-    s32 unk2;
+    f32 unk2;
     s32 unk3;
 
     func_800EE7F8(&unk1, &self->kazfly->unk10);
